+++ src/core/requestHandlers/songRequest.ts
@@ -1,7 +1,7 @@
 import type { User, VoiceBasedChannel } from 'discord.js';
 import { activePlayers } from 'src/Bot';
 import { pushSongs, updateVoiceChannel } from 'src/database/queries/guilds/update';
-import { extractSongData, extractVoiceChannelData } from 'src/utils/utils';
+import { extractSongData, extractVoiceChannelData } from 'src/utils/songUtils';
 
 import { sendEmbed, sendErrorEmbed } from '../messages';
 import { songPlayer } from '../player';
