+++ src/utils/messageUtils.ts
@@ -1,0 +1,9 @@
+import { secrets } from '../config/secrets';
+import type { MessageFormaterInterface } from './interfaces';
+
+export const messageFormater = (message: string): MessageFormaterInterface => {
+  return {
+    splittedMessage: message.split(' '),
+    command: message.split(' ')[0].slice(secrets.PREFIX.length),
+  };
+};
